backend/requirements.txt
Django>=4.1.7,<4.2
django-environ>=0.10.0,<0.11
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.2.2,<5.3
drf_spectacular>=0.26.1,<0.27
